Move load_weights and save_weights to trainer/utils
